Updated code structures and url mapping.
- API 통신 컨트롤러와 일반 MVC 컨트롤러를 별도의 디렉토리에 분리.
- 목적에 맞도록 핸들러 맵핑 이동.
- id가 필수적이지 않은(다시 입력받는 상호작용이 가능한) 핸들러 메서드의 경우 파라미터로 주어지지 않았을 때 400 에러를 발생시키지 않도록 변경.
- 바뀐 URL에 따라 템플릿도 수정.
