+++ src/main/java/com/programmers/coffeeorder/controller/DeliveryController.java
@@ -3,9 +3,11 @@ package com.programmers.coffeeorder.controller;
 import com.programmers.coffeeorder.controller.bind.CoffeeOrderDeliveryRequest;
 import com.programmers.coffeeorder.entity.delivery.CoffeeOrderDelivery;
 import com.programmers.coffeeorder.entity.delivery.DeliveryStatus;
+import com.programmers.coffeeorder.entity.order.CoffeeOrder;
 import com.programmers.coffeeorder.service.delivery.CoffeeDeliveryService;
 import com.programmers.coffeeorder.service.order.CoffeeOrderService;
 import lombok.RequiredArgsConstructor;
+import org.springframework.format.annotation.DateTimeFormat;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -13,7 +15,10 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 
+import java.time.LocalDate;
 import java.time.LocalDateTime;
+import java.util.List;
+import java.util.Map;
 
 @Controller
 @RequestMapping("/delivery")
@@ -29,22 +34,26 @@ public class DeliveryController {
 
 
     @GetMapping
-    public String readDeliveryInfo(@RequestParam("id") long id,
+    public String readDeliveryInfo(@RequestParam(value = "id", required = false) Long id,
                                    Model model) {
-        model.addAttribute("id", id);
-        coffeeDeliveryService.readCoffeeOrderDelivery(id).ifPresentOrElse(
-                delivery -> model.addAttribute("delivery", delivery),
-                () -> model.addAttribute(ERROR_MESSAGE_COMPONENT, "Delivery with given id not found."));
+        if(id != null) {
+            model.addAttribute("id", id);
+            coffeeDeliveryService.readCoffeeOrderDelivery(id).ifPresentOrElse(
+                    delivery -> model.addAttribute("delivery", delivery),
+                    () -> model.addAttribute(ERROR_MESSAGE_COMPONENT, "Delivery with given id not found."));
+        }
         return DELIVERY_READ_TEMPLATE;
     }
 
     @GetMapping("/create")
-    public String requestDeliveryCreate(@RequestParam("id") long id,
+    public String requestDeliveryCreate(@RequestParam(value = "id", required = false) Long id,
                                         Model model) {
-        model.addAttribute("id", id);
-        coffeeOrderService.readOrder(id).ifPresentOrElse(
-                order -> model.addAttribute("order", order),
-                () -> model.addAttribute(ERROR_MESSAGE_COMPONENT, "Order with given id not found."));
+        if(id != null) {
+            model.addAttribute("id", id);
+            coffeeOrderService.readOrder(id).ifPresentOrElse(
+                    order -> model.addAttribute("order", order),
+                    () -> model.addAttribute(ERROR_MESSAGE_COMPONENT, "Order with given id not found."));
+        }
         return DELIVERY_REQUEST_TEMPLATE;
     }
 
@@ -78,4 +87,14 @@ public class DeliveryController {
         }
     }
 
+    @GetMapping("/appointed")
+    public String getCoffeeDeliveryReservations(@RequestParam(value = "date", required = false)
+                                                @DateTimeFormat(iso = DateTimeFormat.ISO.DATE) LocalDate date,
+                                                Model model) {
+        if (date == null) date = LocalDate.now();
+        Map<String, List<CoffeeOrder.DTO>> deliveries = coffeeDeliveryService.listAppointedDeliveries(date);
+        model.addAttribute("deliveries", deliveries);
+        model.addAttribute("date", date);
+        return "deliveries/deliveries-scheduled";
+    }
 }
