+++ src/main/java/com/programmers/coffeeorder/controller/OrderManagerController.java
@@ -20,17 +20,17 @@ import java.util.Map;
 import java.util.stream.Collectors;
 
 @Controller
-@RequestMapping("/manage")
+@RequestMapping("/order")
 @RequiredArgsConstructor
-public class ManagerController {
+public class OrderManagerController {
 
     private final CoffeeOrderService coffeeOrderService;
     private final CoffeeDeliveryService coffeeDeliveryService;
     private final CoffeeProductService coffeeProductService;
 
-    private static final String REDIRECT_TO_ORDER_LIST = "redirect:/manage/orders";
+    private static final String REDIRECT_TO_ORDER_LIST = "redirect:/order";
 
-    @GetMapping("/orders")
+    @GetMapping
     // https://javacan.tistory.com/468 use custom resolver?
     public String getCoffeeOrders(@RequestParam(value = "from", required = false)
                                   @DateTimeFormat(iso = DateTimeFormat.ISO.DATE_TIME) LocalDateTime from, // https://www.baeldung.com/spring-date-parameters
@@ -46,23 +46,23 @@ public class ManagerController {
         return "manage/orders";
     }
 
-    @GetMapping("/orders/accept")
+    @GetMapping("/accept")
     public String acceptCoffeeOrder(@RequestParam("id") Long id) {
         if (id != null) coffeeOrderService.acceptOrder(id);
-        return REDIRECT_TO_ORDER_LIST;
+        return REDIRECT_TO_ORDER_LIST + String.format("#order-%d", id);
     }
 
-    @GetMapping("/orders/cancel")
+    @GetMapping("/cancel")
     public String cancelCoffeeOrder(@RequestParam(name = "id") Long id) {
         if (id != null) coffeeOrderService.cancelOrder(id);
-        return REDIRECT_TO_ORDER_LIST;
+        return REDIRECT_TO_ORDER_LIST + String.format("#order-%d", id);
     }
 
-    @GetMapping("/orders/update")
-    public String requestUpdateCoffeeOrder(@RequestParam("id") Long id,
+    @GetMapping("/update")
+    public String requestUpdateCoffeeOrder(@RequestParam(value = "id", required = false) Long id,
                                            Model model) {
-        if (id == null) return REDIRECT_TO_ORDER_LIST;
         model.addAttribute("id", id);
+        if (id == null) return "manage/update-order";
 
         coffeeOrderService.readOrder(id).ifPresentOrElse(
                 order -> {
@@ -81,28 +81,12 @@ public class ManagerController {
         return "manage/update-order";
     }
 
-    @PostMapping("/orders/update")
+    @PostMapping("/update")
     public String submitUpdateCoffeeOrder(CoffeeOrderUpdate update) {
         CoffeeOrder updatedCoffeeOrder = new CoffeeOrder(update);
         coffeeOrderService.updateOrderInfo(update.getId(), updatedCoffeeOrder);
         coffeeOrderService.updateOrderItemsQuantity(update.getId(), update.getOrderItems());
-        return "redirect:/manage/orders/update?id=" + update.getId();
-    }
-
-    @GetMapping("/deliveries")
-    public String getCoffeeDeliveryReservations(@RequestParam(value = "date", required = false)
-                                                @DateTimeFormat(iso = DateTimeFormat.ISO.DATE) LocalDate date,
-                                                Model model) {
-        if (date == null) date = LocalDate.now();
-        Map<String, List<CoffeeOrder.DTO>> deliveries = coffeeDeliveryService.listAppointedDeliveries(date);
-        model.addAttribute("deliveries", deliveries);
-        model.addAttribute("date", date);
-        return "manage/deliveries-scheduled";
-    }
-
-    @ExceptionHandler(IllegalArgumentException.class)
-    public String handleIllegalArgument() {
-        return REDIRECT_TO_ORDER_LIST;
+        return "redirect:/order/update?id=" + update.getId();
     }
 
 }
